refactor(main): give each window its own file, and the mac App menu to the template
`windowManager.ts` held both windows, and the two creation functions had
grown a shared middle: the same title-bar chrome, the same dark ground, the
same activate-and-act, the same console forwarder, the same geometry
persistence. They matched because two copies happened to agree -- a
resemblance that lasts until someone edits one of them.

`windows/` now has one file per concern. `windowChrome.ts` owns the dressing
both windows wear, so the shared options are shared rather than duplicated
and each window's own file states only what makes it that window: the main
window is the one that registers the IPC surface and watches the renderer
process; the Rendering window is the one placed over its parent and clamped
into the work area. `windowState.ts` (remembering where a window was) and
`closeFunnel.ts` (letting the renderer answer before a window closes) each
serve both. `windowManager.ts` stays as the name the rest of main asks for
them by.

The macOS Application menu moves the other way, out of the Menu builder and
into `shared/menuTemplate.ts` as `macAppMenuGroup(appName)`. It was a
literal nested inside `buildAndSetMenu`, which is why the group-level
`darwinOnly` field next to `APP_MENU` had been reduced to a documented
no-op, and why its test had to read `menu.ts` as text and grep for item ids.
Both menus are data in one file now, and the test asserts on the group.
